added hicrop tests
