+++ isis/src/mro/apps/hicrop/hicrop.h
@@ -1,0 +1,12 @@
+#ifndef hicrop_h 
+#define hicrop_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void hicrop(Cube *cube, UserInterface &ui, Pvl *log=nullptr);
+  extern void hicrop(UserInterface &ui, Pvl *log=nullptr);
+}
+
+#endif
