+++ isis/tests/Fixtures.cpp
@@ -1,3 +1,5 @@
+#include <QTextStream>
+
 #include "Fixtures.h"
 #include "LineManager.h"
 #include "SpecialPixel.h"
@@ -24,7 +26,6 @@ namespace Isis {
       }
       testCube->write(line);
     }
-
   }
 
   void SmallCube::TearDown() {
@@ -106,16 +107,18 @@ namespace Isis {
 
 
   void DefaultCube::TearDown() {
+    std::cout << "in tear down" << std::endl;
     if (testCube->isOpen()) {
       testCube->close();
     }
-
+    std::cout << "closing proj" << std::endl;
     if (projTestCube->isOpen()) {
       projTestCube->close();
     }
-
+    std::cout << "before delete" << std::endl;
     delete testCube;
     delete projTestCube;
+    std::cout << "after delete" << std::endl;
   }
 
   void LineScannerCube::SetUp() {
@@ -221,6 +224,7 @@ namespace Isis {
     delete threeImageOverlapFile;
     delete twoImageOverlapFile;
   }
+  
 
   void MroCube::setInstrument(QString ikid, QString instrumentId, QString spacecraftName) {
     PvlGroup &kernels = testCube->label()->findObject("IsisCube").findGroup("Kernels");
@@ -232,12 +236,12 @@ namespace Isis {
         SpacecraftName              = "MARS RECONNAISSANCE ORBITER"
         InstrumentId                = HIRISE
         TargetName                  = Mars
-        StartTime                   = 2006-11-08T04:49:13.968
-        StopTime                    = 2006-11-08T04:49:17.771
-        ObservationStartCount       = 847428572:42722
-        SpacecraftClockStartCount   = 847428572:52459
-        SpacecraftClockStopCount    = 847428576:39516
-        ReadoutStartCount           = 847428727:63203
+        StartTime                   = 2008-05-17T09:37:24.7300819
+        StopTime                    = 2008-05-17T09:37:31.0666673
+        ObservationStartCount       = 895484264:44383
+        SpacecraftClockStartCount   = 895484264:57342
+        SpacecraftClockStopCount    = 895484272:12777
+        ReadoutStartCount           = 895484659:31935
         CalibrationStartTime        = 2006-11-08T04:49:13.952
         CalibrationStartCount       = 847428572:51413
         AnalogPowerStartTime        = 2006-11-08T04:48:34.478
@@ -310,7 +314,6 @@ namespace Isis {
     PvlGroup newInstGroup; 
     iss >> newInstGroup; 
     
-
     newInstGroup.findKeyword("InstrumentId").setValue(instrumentId);
     newInstGroup.findKeyword("SpacecraftName").setValue(spacecraftName);
 
@@ -347,11 +350,32 @@ namespace Isis {
         testCube->write(line);
     }
     testCube->reopen("rw");
- 
+  
     // need to remove old camera pointer 
     delete testCube;
     // This is now a MRO cube
+
     testCube = new Cube(fileName, "rw");
+
+    // create a jitter file 
+    QString jitter = R"(# Sample                 Line                   ET
+-0.18     -0.07     264289109.96933
+-0.11     -0.04     264289109.97
+-0.05     -0.02     264289109.98
+1.5     0.6     264289110.06
+    )"; 
+    
+    jitterPath = tempDir.path() + "/jitter.txt"; 
+    QFile jitterFile(jitterPath); 
+    
+    if (jitterFile.open(QIODevice::WriteOnly)) {
+      QTextStream out(&jitterFile); 
+      out << jitter;
+      jitterFile.close(); 
+    }
+    else { 
+      FAIL() << "Failed to create Jitter file" << std::endl;
+    }
   }
