+++ isis/tests/Fixtures.h
@@ -113,6 +113,11 @@ namespace Isis {
 
   class MroCube : public DefaultCube {
     protected:
+      QString ckPath = "data/mroKernels/mroCK.bc";
+      QString sclkPath = "data/mroKernels/mroSCLK.tsc";
+      QString lskPath = "data/mroKernels/mroLSK.tls";
+      QString jitterPath; 
+
       void setInstrument(QString ikid, QString instrumentId, QString spacecraftName); 
   };
